'debug'

+++ database/migrations/2022_05_10_073022_create_wallets_table.php
@@ -18,8 +18,8 @@ class CreateWalletsTable extends Migration
             $table->timestamps();
             $table->string('title');
             $table->string('comment')->nullable();
-            $table->Integer('priceBuy');
-            $table->Integer('quantity');
+            $table->float('priceBuy');
+            $table->float('quantity');
             $table->string('device');
             $table->date('date');
             $table->unsignedBigInteger('id_user');
